backend/requirements.txt
Django>=5.1.4
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
xlwt>=1.3.0  # For Excel export
openpyxl>=3.1.2  # For Excel export 